// src/app/api/swagger/models/api-tag-participant-offer-rule-type-enum.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiTagParticipantOfferRuleTypeEnum {
  ParticipantOffer = 1,
  ParticipantOfferCategory = 2,
  ParticipantOfferOkpd2 = 3,
  ParticipantOfferTenant = 4,
  ParticipantId = 5,
  ParticipantHomeRegion = 6,
  PriceList = 7,
  HasUinTru = 8,
  ParticipantOfferProductName = 9
}